feat(scripts): Modularize README Generator script
