+++ scripts/generate_readme/README.md
@@ -1,0 +1,88 @@
+# Generate README Module
+
+A modular tool for automatically generating README documentation for Kubeflow Pipelines components and pipelines.
+
+## Structure
+
+```
+generate_readme/
+├── __init__.py           # Package initialization
+├── __main__.py           # Entry point for module execution
+├── cli.py                # Command-line interface and argument parsing
+├── constants.py          # Shared constants and logger configuration
+├── content_generator.py  # README content generation logic
+├── generator.py          # Main README generator orchestration
+└── metadata_parser.py    # Metadata extraction from KFP components/pipelines
+```
+
+## Usage
+
+Run from the project root directory:
+
+```bash
+# Generate README for a component
+python -m scripts.generate_readme --component components/my_component
+
+# Generate README for a pipeline
+python -m scripts.generate_readme --pipeline pipelines/my_pipeline
+
+# With additional options
+python -m scripts.generate_readme --component components/my_component --verbose --overwrite
+
+# Or with uv
+uv run python -m scripts.generate_readme --component components/my_component
+```
+
+## Features
+
+- **Automatic metadata extraction**: Parses Python functions decorated with `@dsl.component` or `@dsl.pipeline`
+- **Google-style docstring parsing**: Extracts parameter descriptions and return values
+- **KEP-913 compliance**: Generates READMEs following the standardized template
+- **Custom content preservation**: Preserves user-added content after the `<!-- custom-content -->` marker
+- **Type annotation support**: Handles complex type annotations including Optional, Union, and generics
+- **Component-specific usage examples**: Generates usage examples for components (not pipelines)
+
+## Module Components
+
+### metadata_parser.py
+
+- `MetadataParser`: Base class with shared parsing utilities
+- `ComponentMetadataParser`: Extracts metadata from `@dsl.component` functions
+- `PipelineMetadataParser`: Extracts metadata from `@dsl.pipeline` functions
+
+### content_generator.py
+
+- `ReadmeContentGenerator`: Generates formatted README sections from extracted metadata
+
+### generator.py
+
+- `ReadmeGenerator`: Orchestrates the README generation process
+
+### cli.py
+
+- `validate_component_directory()`: Validates component directory structure
+- `validate_pipeline_directory()`: Validates pipeline directory structure
+- `parse_arguments()`: Parses command-line arguments
+- `main()`: Entry point for CLI execution
+
+## Custom Content
+
+Users can add custom sections to their READMEs that will be preserved across regenerations:
+
+1. Add the marker `<!-- custom-content -->` at the desired location
+2. Write custom content below the marker
+3. The content will be preserved when regenerating the README
+
+Example:
+
+```markdown
+## Metadata 🗂️
+...
+
+<!-- custom-content -->
+
+## Additional Examples
+
+Custom examples that won't be overwritten...
+```
+
